Fix download redirects with functions
This resolves an issue introduced in the redirect prep for v3 that
broke our download redirects.

The problem is that v3 releases are `kubebuilder_${os}_${arch}`, whereas
v2 & v1 releases are `kubebuilder_${version}_${os}_${arch}.tar.gz`.

Since netlify can't handle wildcards that are part of a path component
(like `/2.:minorversion`) instead of a whole one (`:version`), we can't
select on major version directly in our redirects.

Instead, we introduce a function that handles this logic for us, and
then use netlify 200-pseudo-redirects to "mount the function on the
downloads part of the releases endpoint.